Renamed app components
