+++ front/src/components/App/StatisticsArea/Statistics.tsx
